[DOC]/23: Changes in Swiper.vue in order to get Swiper.stories.tsx working. Swiper.stories.tsx is alive and running
